+++ lib/prefixed_ids.rb
@@ -74,7 +74,7 @@ module PrefixedIds
 
     class_methods do
       def find(*ids)
-        super(*ids.map { |id| _prefix_id.decode(id) })
+        super(*ids.map { |id| _prefix_id.decode(id, fallback: true) })
       end
     end
   end